Fixed code suggestions

+++ Maya/Exporter/BabylonExporter.Mesh.cs
@@ -1019,15 +1019,15 @@ namespace Maya2Babylon
                 // decreasing sort
                 OrderByDescending(ref weightByInfluenceIndex);
 
-                long bonesCount = indexByNodeName.Count; // number total of bones/influences for the mesh
+                int bonesCount = indexByNodeName.Count; // number total of bones/influences for the mesh
                 float weight0 = 0;
                 float weight1 = 0;
                 float weight2 = 0;
                 float weight3 = 0;
-                long bone0 = 0;
-                long bone1 = 0;
-                long bone2 = 0;
-                long bone3 = 0;
+                int bone0 = 0;
+                int bone1 = 0;
+                int bone2 = 0;
+                int bone3 = 0;
                 int nbBones = weightByInfluenceIndex.Count; // number of bones/influences for this vertex
 
                 if (nbBones == 0)
@@ -1062,7 +1062,7 @@ namespace Maya2Babylon
 
                 float[] weights = { weight0, weight1, weight2, weight3 };
                 vertex.Weights = weights;
-                int[] boneIndexes = { (int)bone0, (int)bone1, (int)bone2, (int)bone3 };
+                int[] boneIndexes = { bone0, bone1, bone2, bone3 };
                 vertex.BonesIndices = boneIndexes;
 
                 if (nbBones > 4)
@@ -1093,7 +1093,7 @@ namespace Maya2Babylon
 
                     float[] weightsExtra = { weight0, weight1, weight2, weight3 };
                     vertex.WeightsExtra = weightsExtra;
-                    int[] boneIndexesExtra = { (int)bone0, (int)bone1, (int)bone2, (int)bone3 };
+                    int[] boneIndexesExtra = { bone0, bone1, bone2, bone3 };
                     vertex.BonesIndicesExtra = boneIndexesExtra;
                 }
             }
